+++ tests/integration/other/components/mutually_exclusive/schema_urls.py
