+++ tests/integration/other/components/mutually_exclusive/test_date_single_checkbox_override.py
@@ -1,4 +1,4 @@
-from tests.integration.components.mutually_exclusive.schema_urls import (
+from tests.integration.other.components.mutually_exclusive.schema_urls import (
     MUTUALLY_EXCLUSIVE_DAY_MONTH_YEAR_DATE,
 )
 from tests.integration.integration_test_case import IntegrationTestCase
