+++ tests/integration/other/components/mutually_exclusive/test_textfield_single_checkbox_override.py
@@ -1,55 +1,55 @@
-from tests.integration.components.mutually_exclusive.schema_urls import (
-    MUTUALLY_EXCLUSIVE_TEXTAREA,
+from tests.integration.other.components.mutually_exclusive.schema_urls import (
+    MUTUALLY_EXCLUSIVE_TEXTFIELD,
 )
 from tests.integration.integration_test_case import IntegrationTestCase
 
 
-class TestTextAreaSingleCheckboxOverride(IntegrationTestCase):
+class TestTextFieldSingleCheckboxOverride(IntegrationTestCase):
     """
-    Tests to ensure that the server-side validation for mutually exclusive textarea
+    Tests to ensure that the server-side validation for mutually exclusive textfield
     with single checkbox override function as expected. These tests emulate the non-JS version.
     """
 
     def setUp(self):
         super().setUp()
         self.launchSurvey("test_mutually_exclusive")
-        self.get(MUTUALLY_EXCLUSIVE_TEXTAREA)
+        self.get(MUTUALLY_EXCLUSIVE_TEXTFIELD)
 
     def test_non_exclusive_answer(self):
         # When
-        self.post({"textarea-answer": "John Doe"})
+        self.post({"textfield-answer": "John Doe"})
 
         # Then
-        self.assertInUrl("/sections/mutually-exclusive-textarea-section/")
+        self.assertInUrl("/sections/mutually-exclusive-textfield-section/")
         self.assertInBody("John Doe")
 
     def test_exclusive_answer(self):
         # When
         self.post(
             {
-                "textarea-answer": "",
-                "textarea-exclusive-answer": ["I prefer not to say"],
+                "textfield-answer": "",
+                "textfield-exclusive-answer": ["I prefer not to say"],
             }
         )
 
         # Then
-        self.assertInUrl("/sections/mutually-exclusive-textarea-section/")
+        self.assertInUrl("/sections/mutually-exclusive-textfield-section/")
         self.assertInBody("I prefer not to say")
 
     def test_optional_exclusive_question(self):
         # When
-        self.post({"textarea-answer": ""})
+        self.post({"textfield-answer": ""})
 
         # Then
-        self.assertInUrl("/sections/mutually-exclusive-textarea-section/")
+        self.assertInUrl("/sections/mutually-exclusive-textfield-section/")
         self.assertInBody("No answer provided")
 
     def test_invalid_exclusive_answers(self):
         # When
         self.post(
             {
-                "textarea-answer": "John Doe",
-                "textarea-exclusive-answer": ["I prefer not to say"],
+                "textfield-answer": "John Doe",
+                "textfield-exclusive-answer": ["I prefer not to say"],
             }
         )
