+++ tests/integration/other/components/test_cookies_banner_content.py
