+++ tests/integration/other/components/test_render_dropdown_widget.py
