+++ tests/integration/other/components/test_render_percentage_widget.py
