+++ tests/integration/other/session/test_sign_out_and_exit.py
