+++ tests/integration/other/test_application_variables.py
