+++ tests/integration/other/test_application_variables_negative.py
