+++ tests/integration/routes/routing/test_routing_case_insensitive.py
